30706 moved seq-workflow components into a new folder
