+++ packages/dina-ui/components/seqdb/seq-workflow/seq-batch-select-coordinats-step/DraggableSeqReactionBox.tsx
